+++ src/widgets/userListSection/model/types.ts
@@ -1,0 +1,3 @@
+export type UserListSectionFormData = {
+  tier: string;
+};
